added a hook for tag table row options
