+++ frontend/dialob-dashboard-material/src/util/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './useFetchAuth';
 export * from './helperFunctions';
+export * from './useFormTagOptions';
